Rename screen_recording library as screen_recording_controls
to better reflect the intended purpose of controlling running devices'
screen recordings as opposed to actually implementing the feature.
