+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -146,7 +146,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:config_utils",
         "//cuttlefish/host/libs/config:cuttlefish_config",
-        "//cuttlefish/host/libs/screen_recording",
+        "//cuttlefish/host/libs/screen_recording_controls",
         "//libbase",
         "@fmt",
         "@jsoncpp",
