+++ base/cvd/cuttlefish/host/commands/record_cvd/BUILD.bazel
@@ -12,7 +12,7 @@ cf_cc_binary(
     deps = [
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/libs/config:cuttlefish_config",
-        "//cuttlefish/host/libs/screen_recording",
+        "//cuttlefish/host/libs/screen_recording_controls",
         "//libbase",
         "@gflags",
     ],
