+++ base/cvd/cuttlefish/host/commands/screen_recording_server/BUILD.bazel
@@ -39,7 +39,7 @@ cf_cc_binary(
         "//cuttlefish/host/libs/command_util:libcuttlefish_run_cvd_proto",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:logging",
-        "//cuttlefish/host/libs/screen_recording",
+        "//cuttlefish/host/libs/screen_recording_controls",
         "//libbase",
         "@gflags",
         "@grpc",
